+++ src/newgrf_railtype.cpp
@@ -32,7 +32,7 @@ static void RailTypeSetTriggers(const ResolverObject *object, int triggers)
 {
 }
 
-static uint32 RailTypeGetVariable(const ResolverObject *object, byte variable, byte parameter, bool *available)
+static uint32 RailTypeGetVariable(const ResolverObject *object, byte variable, uint32 parameter, bool *available)
 {
 	TileIndex tile = object->u.routes.tile;
